+++ cmake/DetectLibraries.cmake
@@ -1,6 +1,17 @@
 # -------------------------------------------------------------------------------------------------- #
+# вырабатываем и подключаем файл с ресурсами библиотеки
+if( WIN32 )
+  configure_file( ${CMAKE_CURRENT_SOURCE_DIR}/libakrypt.rc.in ${CMAKE_CURRENT_BINARY_DIR}/libakrypt.rc @ONLY )
+  configure_file( ${CMAKE_CURRENT_SOURCE_DIR}/libakrypt-base.rc.in ${CMAKE_CURRENT_BINARY_DIR}/libakrypt-base.rc @ONLY )
+  set( AKRYPT_SOURCES ${AKRYPT_SOURCES} ${CMAKE_CURRENT_BINARY_DIR}/libakrypt.rc )
+  set( AKBASE_SOURCES ${AKBASE_SOURCES} ${CMAKE_CURRENT_BINARY_DIR}/libakrypt-base.rc )
+  set( CMAKE_BUILD_TYPE "Release" )
+  message("-- Generation of ${CMAKE_CURRENT_BINARY_DIR}/libakrypt.rc is done")
+endif()
+
+# -------------------------------------------------------------------------------------------------- #
 # поиск gmp
-if( LIBAKRYPT_GMP_TESTS )
+if( AK_TESTS_GMP )
 
   find_library( LIBGMP gmp )
   if( LIBGMP )
@@ -87,12 +98,4 @@ else()
       set( CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DAK_HAVE_PTHREAD_H" )
     endif()
   endif()
-endif()
-
-# -------------------------------------------------------------------------------------------------- #
-# вырабатываем и подключаем файл с ресурсами библиотеки
-if( WIN32 )
-  configure_file( ${CMAKE_CURRENT_SOURCE_DIR}/libakrypt.rc.in ${CMAKE_CURRENT_SOURCE_DIR}/libakrypt.rc @ONLY )
-  set( AKRYPT_SOURCES ${AKRYPT_SOURCES} libakrypt.rc )
-  set( CMAKE_BUILD_TYPE "Release" )
 endif()