+++ source/ak_options.c
@@ -139,11 +139,11 @@
 /* ----------------------------------------------------------------------------------------------- */
  void ak_libakrypt_log_options( void )
 {
- /* выводим сообщение об установленных параметрах библиотеки */
     size_t i = 0;
-    ak_error_message_fmt( ak_error_ok, __func__, "libakrypt version: %s", ak_libakrypt_version( ));
-   /* далее мы пропускаем вывод информации об архитектуре,
-    поскольку она будет далее тестироваться отдельно     */
+
+   /* выводим сообщение об установленных параметрах библиотеки */
+   /* мы пропускаем вывод информации об архитектуре,
+      поскольку она будет далее тестироваться отдельно         */
     for( i = 1; i < ak_libakrypt_options_count(); i++ ) {
        switch( options[i].value ) {
 
@@ -350,8 +350,10 @@
   if( user != NULL ) return 0;
   if( strncmp( section, "libakrypt", 9 ) != 0 ) return 0;
   if( strncmp( name, "certificate_repository", 25 ) == 0 ) {
-    if( ak_certificate_set_repository( valstr ) != ak_error_ok ) return 0;
-     else return 1;
+    if( ak_certificate_set_repository( valstr ) != ak_error_ok ) ak_error_set_value( ak_error_ok );
+   /* если каталог с доверенными сертификатами не существует,
+      мы выдаём сообщение об ошибке и продолжаем работу программы */
+    return 1;
   }
 
  /* теперь детальный разбор каждой опции */
@@ -393,12 +395,12 @@
  if( ak_file_open_to_read( &fd, name ) == ak_error_ok ) {
    ak_file_close( &fd );
    if(( error = ak_ini_parse( name, ak_libakrypt_load_option_from_file, NULL )) == ak_error_ok ) {
-     if( ak_libakrypt_get_option_by_name( "log_level" ) > ak_log_none )
-       ak_error_message_fmt( ak_error_ok, "ak_libakrypt_log_options",
+     if( ak_log_get_level() > ak_log_none ) ak_error_message_fmt( ak_error_ok, __func__,
                                             "all options have been read from the %s file", name );
      return ak_true;
    } else {
-       ak_error_message_fmt( error, __func__, "file %s exists, but contains invalid data", name );
+       ak_error_message_fmt( ak_error_wrong_option, __func__,
+                           "file %s exists, but contains invalid data in line: %d", name, error );
        return ak_false;
      }
  }
@@ -413,12 +415,12 @@
  if( ak_file_open_to_read( &fd, name ) == ak_error_ok ) {
    ak_file_close( &fd );
    if(( error = ak_ini_parse( name, ak_libakrypt_load_option_from_file, NULL )) == ak_error_ok ) {
-     if( ak_libakrypt_get_option_by_name( "log_level" ) > ak_log_none )
-       ak_error_message_fmt( ak_error_ok, "ak_libakrypt_log_options",
+     if( ak_log_get_level() > ak_log_none ) ak_error_message_fmt( ak_error_ok, __func__,
                                              "all options have been read from the %s file", name );
      return ak_true;
    } else {
-       ak_error_message_fmt( error, __func__, "wrong options reading from %s file", name );
+       ak_error_message_fmt( ak_error_wrong_option, __func__,
+                           "file %s exists, but contains invalid data in line: %d", name, error );
        return ak_false;
      }
  } else ak_error_message( ak_error_access_file, __func__,